Split old AVX2 implementations as a branch

+++ src/Titanium_CCA_hi/pack.c
@@ -18,7 +18,7 @@ static inline uint32_t con_sub(uint32_t x) /* conditional subtraction of q */
 }
 
 /* convert a polynomial to a binary string */
-void poly_encode(unsigned char *b, const uint32_t *p, uint32_t len)
+void poly_encode(unsigned char *b, const uint64_t *p, uint32_t len)
 {
 	uint32_t i;
 	unsigned char *bb;
@@ -41,7 +41,7 @@ void poly_encode(unsigned char *b, const uint32_t *p, uint32_t len)
 }
 
 /* convert a binary string to a polynomial */
-void poly_decode(uint32_t *p, const unsigned char *b, uint32_t len)
+void poly_decode(uint64_t *p, const unsigned char *b, uint32_t len)
 {
 	uint32_t i;
 	unsigned char *bb;
@@ -51,13 +51,13 @@ void poly_decode(uint32_t *p, const unsigned char *b, uint32_t len)
 	{
 		bb = b + (i / 2) * Q_BITS_PACK;
 
-		p[i] = ((uint32_t)bb[0]) | (((uint32_t)bb[1]) << 8) | ((((uint32_t)bb[2]) & 0x0f) << 16);
-		p[i + 1] = (((uint32_t)bb[2]) >> 4) | (((uint32_t)bb[3]) << 4) | (((uint32_t)bb[4]) << 12);
+		p[i] = ((uint64_t)bb[0]) | (((uint64_t)bb[1]) << 8) | ((((uint64_t)bb[2]) & 0x0f) << 16);
+		p[i + 1] = (((uint64_t)bb[2]) >> 4) | (((uint64_t)bb[3]) << 4) | (((uint64_t)bb[4]) << 12);
 	}
 }
 
 /* convert a polynomial to a binary string (with compression) */
-void poly_encode_c2(unsigned char *b, const uint32_t *p, uint32_t len)
+void poly_encode_c2(unsigned char *b, const uint64_t *p, uint32_t len)
 {
 	uint32_t i;
 
@@ -69,14 +69,14 @@ void poly_encode_c2(unsigned char *b, const uint32_t *p, uint32_t len)
 }
 
 /* convert a binary string to a polynomial (with compression) */
-void poly_decode_c2(uint32_t *p, const unsigned char *b, uint32_t len)
+void poly_decode_c2(uint64_t *p, const unsigned char *b, uint32_t len)
 {
 	uint32_t i;
 	
 	for (i = 0; i < len; i++)
 	{
 		/* shift the compressed coordinates back */
-		p[i] = ((uint32_t)(LOAD_C2(b + i * C2_COMPRESSION_BYTE))) << C2_COMPRESSION_BITS;
+		p[i] = ((uint64_t)(LOAD_C2(b + i * C2_COMPRESSION_BYTE))) << C2_COMPRESSION_BITS;
 	}
 }
