+++ src/Titanium_CCA_hi/pack.h
@@ -10,9 +10,9 @@
 
 #include <stdint.h>
 
-void poly_encode(unsigned char *b, const uint32_t *p, uint32_t len);
-void poly_decode(uint32_t *p, const unsigned char *b, uint32_t len);
-void poly_encode_c2(unsigned char *b, const uint32_t *p, uint32_t len);
-void poly_decode_c2(uint32_t *p, const unsigned char *b, uint32_t len);
+void poly_encode(unsigned char *b, const uint64_t *p, uint32_t len);
+void poly_decode(uint64_t *p, const unsigned char *b, uint32_t len);
+void poly_encode_c2(unsigned char *b, const uint64_t *p, uint32_t len);
+void poly_decode_c2(uint64_t *p, const unsigned char *b, uint32_t len);
 
 #endif
