+++ src/Titanium_CPA_toy/pack.c
@@ -18,7 +18,7 @@ static inline uint32_t con_sub(uint32_t x) /* conditional subtraction of q */
 }
 
 /* convert a polynomial to a binary string */
-void poly_encode(unsigned char *b, const uint32_t *p, uint32_t len)
+void poly_encode(unsigned char *b, const uint64_t *p, uint32_t len)
 {
 	uint32_t i;
 	unsigned char *bb;
@@ -47,7 +47,7 @@ void poly_encode(unsigned char *b, const uint32_t *p, uint32_t len)
 }
 
 /* convert a binary string to a polynomial */
-void poly_decode(uint32_t *p, const unsigned char *b, uint32_t len)
+void poly_decode(uint64_t *p, const unsigned char *b, uint32_t len)
 {
 	uint32_t i;
 	unsigned char *bb;
@@ -57,15 +57,15 @@ void poly_decode(uint32_t *p, const unsigned char *b, uint32_t len)
 	{
 		bb = b + (i / 4) * Q_BITS_PACK;
 
-		p[i] = ((uint32_t)bb[0]) | (((uint32_t)bb[1]) << 8) | ((((uint32_t)bb[2]) & 0x03) << 16);
-		p[i + 1] = (((uint32_t)bb[2]) >> 2) | (((uint32_t)bb[3]) << 6) | ((((uint32_t)bb[4]) & 0x0f) << 14);
-		p[i + 2] = (((uint32_t)bb[4]) >> 4) | (((uint32_t)bb[5]) << 4) | ((((uint32_t)bb[6]) & 0x3f) << 12);
-		p[i + 3] = (((uint32_t)bb[6]) >> 6) | (((uint32_t)bb[7]) << 2) | (((uint32_t)bb[8]) << 10);		
+		p[i] = ((uint64_t)bb[0]) | (((uint64_t)bb[1]) << 8) | ((((uint64_t)bb[2]) & 0x03) << 16);
+		p[i + 1] = (((uint64_t)bb[2]) >> 2) | (((uint64_t)bb[3]) << 6) | ((((uint64_t)bb[4]) & 0x0f) << 14);
+		p[i + 2] = (((uint64_t)bb[4]) >> 4) | (((uint64_t)bb[5]) << 4) | ((((uint64_t)bb[6]) & 0x3f) << 12);
+		p[i + 3] = (((uint64_t)bb[6]) >> 6) | (((uint64_t)bb[7]) << 2) | (((uint64_t)bb[8]) << 10);		
 	}
 }
 
 /* convert a polynomial to a binary string (with compression) */
-void poly_encode_c2(unsigned char *b, const uint32_t *p, uint32_t len)
+void poly_encode_c2(unsigned char *b, const uint64_t *p, uint32_t len)
 {
 	uint32_t i;
 
@@ -77,14 +77,14 @@ void poly_encode_c2(unsigned char *b, const uint32_t *p, uint32_t len)
 }
 
 /* convert a binary string to a polynomial (with compression) */
-void poly_decode_c2(uint32_t *p, const unsigned char *b, uint32_t len)
+void poly_decode_c2(uint64_t *p, const unsigned char *b, uint32_t len)
 {
 	uint32_t i;
 	
 	for (i = 0; i < len; i++)
 	{
 		/* shift the compressed coordinates back */
-		p[i] = ((uint32_t)(LOAD_C2(b + i * C2_COMPRESSION_BYTE))) << C2_COMPRESSION_BITS;
+		p[i] = ((uint64_t)(LOAD_C2(b + i * C2_COMPRESSION_BYTE))) << C2_COMPRESSION_BITS;
 	}
 }
